+++ CHANGELOG.md
@@ -6,6 +6,11 @@ SPDX-License-Identifier: AGPL-3.0-only
 
 # Changelog
 
+## Version 0.8.2 (development)
+
+- Add check for existing, completed EDS files with same name, and option to `Experiment.run` to overwrite completed or working files.
+- Fix XML protocol parsing bug for increment units with non-qslib EDS files.
+
 ## Version 0.8.1
 
 - Fix bug with BytesIO not being recognized by `Experiment.save_file`.